+++ CHANGES.md
@@ -1,5 +1,6 @@
 Version 0.1
 ----------------------------------------------------------------------
 
+* Renamed from bjsimple to bigjobasync
 * Added installer (setup.py)
 * Added 'username' option to resource description